pyproject.toml
[tool.poetry]
name = "search_vacancies_with_database"
version = "0.1.0"
description = ""
authors = ["Nika Alexandrovna <megrelynika@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
poetry-core = "^2.1.3"
utils = "^1.0.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.16.1"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
